+++ cherab/generomak/plasma/__init__.py
@@ -1,1 +1,2 @@
 from .plasma import get_edge_distributions, get_edge_interpolators, get_edge_plasma
+from .plasma import get_core_distributions, get_core_plasma